# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backend",
        "version": "0.1.0"
    },
    "paths": {
...
        "/ai-function": {
            "get": {
                "tags": [
                    "db"
                ],
                "summary": "Get Ai Functions",
                "operationId": "get_ai_functions",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/AIFunction"
                                    },
                                    "type": "array",
                                    "title": "Response Get Ai Functions Ai Function Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "db"
                ],
                "summary": "Post Ai Function",
                "operationId": "post_ai_function",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AIFunctionRouteInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AIFunction"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "Not authenticated"
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "document already exists"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/ai-function/{ai_function_id}": {
            "get": {
                "tags": [
                    "db"
                ],
                "summary": "Get Ai Function",
                "operationId": "get_ai_function",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "ai_function_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ai Function Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AIFunction"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "db"
                ],
                "summary": "Delete Ai Function",
                "operationId": "delete_ai_function",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "ai_function_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ai Function Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "db"
                ],
                "summary": "Patch Ai Function",
                "operationId": "patch_ai_function",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "ai_function_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ai Function Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AIFunctionPatchInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AIFunction"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "409": {
                        "detail": "document already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AIFunction": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "title": "Name",
                        "examples": [
                            "Summarize Texts"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 1,
                        "title": "Description",
                        "examples": [
                            "Summarizes english texts to a given number_of_sentences."
                        ]
                    },
                    "input_variables": {
                        "items": {
                            "$ref": "#/components/schemas/InputVariable"
                        },
                        "type": "array",
                        "title": "Input Variables",
                        "examples": [
                            [
                                {
                                    "name": "text"
                                },
                                {
                                    "name": "number_of_sentences"
                                }
                            ]
                        ]
                    },
                    "output_schema": {
                        "$ref": "#/components/schemas/JsonSchema-Output",
                        "examples": [
                            {
                                "type": "string"
                            }
                        ]
                    },
                    "assert": {
                        "items": {
                            "$ref": "#/components/schemas/Assertion"
                        },
                        "type": "array",
                        "title": "Assert",
                        "examples": [
                            [
                                {
                                    "type": "icontains",
                                    "value": "the",
                                    "weight": 1
                                },
                                {
                                    "type": "contains",
                                    "value": "thgewgewgewgewge",
                                    "weight": 1
                                }
                            ]
                        ]
                    },
                    "test_cases": {
                        "items": {
                            "$ref": "#/components/schemas/TestCase-Output"
                        },
                        "type": "array",
                        "title": "Test Cases",
                        "examples": [
                            [
                                {
                                    "assert": [
                                        {
                                            "type": "icontains",
                                            "value": "serendipity",
                                            "weight": 5
                                        }
                                    ],
                                    "vars": {
                                        "number_of_sentences": "2",
                                        "text": "The power of serendipity is fascinating. Sometimes, the most unexpected encounters can lead to life-changing experiences. Imagine strolling through a park and stumbling upon a group of musicians, their melodies drawing you in. You pause for a moment, only to realize that this spontaneous moment of joy is exactly what you needed\u2014a break from the routine, a reminder of life's simple pleasures. Serendipity teaches us that not everything needs to be planned. Sometimes, the best moments are the ones that catch us by surprise."
                                    }
                                },
                                {
                                    "assert": [
                                        {
                                            "type": "icontains",
                                            "value": "minimalism",
                                            "weight": 5
                                        }
                                    ],
                                    "vars": {
                                        "number_of_sentences": "2",
                                        "text": "The art of minimalism is more than just decluttering your space\u2014it's about simplifying life. In a world overflowing with choices and distractions, minimalism encourages you to focus on what truly matters. It's about owning fewer things but cherishing each one more deeply. By stripping away the excess, you create room for clarity, intention, and peace. Whether it\u2019s reducing physical possessions or streamlining your daily habits, minimalism can bring a sense of freedom, allowing you to invest time and energy in experiences and relationships that bring genuine joy."
                                    }
                                }
                            ]
                        ]
                    },
                    "_id": {
                        "type": "string",
                        "title": " Id"
                    },
                    "number_of_prompts": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Number Of Prompts"
                    },
                    "username": {
                        "type": "string",
                        "format": "email",
                        "title": "Username"
                    },
                    "creation_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Creation Time"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "input_variables",
                    "output_schema",
                    "assert",
                    "test_cases",
                    "number_of_prompts",
                    "username",
                    "creation_time"
                ],
                "title": "AIFunction"
            },
            "AIFunctionPatchInput": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 40,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1000,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "input_variables": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/InputVariable"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Input Variables"
                    },
                    "output_schema": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/JsonSchema-Input"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "assertions": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Assertion"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assertions"
                    },
                    "test_cases": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/TestCase-Input"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Test Cases"
                    }
                },
                "type": "object",
                "title": "AIFunctionPatchInput"
            },
            "AIFunctionRouteInput": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "title": "Name",
                        "examples": [
                            "Summarize Texts"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 1,
                        "title": "Description",
                        "examples": [
                            "Summarizes english texts to a given number_of_sentences."
                        ]
                    },
                    "input_variables": {
                        "items": {
                            "$ref": "#/components/schemas/InputVariable"
                        },
                        "type": "array",
                        "title": "Input Variables",
                        "examples": [
                            [
                                {
                                    "name": "text"
                                },
                                {
                                    "name": "number_of_sentences"
                                }
                            ]
                        ]
                    },
                    "output_schema": {
                        "$ref": "#/components/schemas/JsonSchema-Input",
                        "examples": [
                            {
                                "type": "string"
                            }
                        ]
                    },
                    "assert": {
                        "items": {
                            "$ref": "#/components/schemas/Assertion"
                        },
                        "type": "array",
                        "title": "Assert",
                        "examples": [
                            [
                                {
                                    "type": "icontains",
                                    "value": "the",
                                    "weight": 1
                                },
                                {
                                    "type": "contains",
                                    "value": "thgewgewgewgewge",
                                    "weight": 1
                                }
                            ]
                        ]
                    },
                    "test_cases": {
                        "items": {
                            "$ref": "#/components/schemas/TestCase-Input"
                        },
                        "type": "array",
                        "title": "Test Cases",
                        "examples": [
                            [
                                {
                                    "assert": [
                                        {
                                            "type": "icontains",
                                            "value": "serendipity",
                                            "weight": 5
                                        }
                                    ],
                                    "vars": {
                                        "number_of_sentences": "2",
                                        "text": "The power of serendipity is fascinating. Sometimes, the most unexpected encounters can lead to life-changing experiences. Imagine strolling through a park and stumbling upon a group of musicians, their melodies drawing you in. You pause for a moment, only to realize that this spontaneous moment of joy is exactly what you needed\u2014a break from the routine, a reminder of life's simple pleasures. Serendipity teaches us that not everything needs to be planned. Sometimes, the best moments are the ones that catch us by surprise."
                                    }
                                },
                                {
                                    "assert": [
                                        {
                                            "type": "icontains",
                                            "value": "minimalism",
                                            "weight": 5
                                        }
                                    ],
                                    "vars": {
                                        "number_of_sentences": "2",
                                        "text": "The art of minimalism is more than just decluttering your space\u2014it's about simplifying life. In a world overflowing with choices and distractions, minimalism encourages you to focus on what truly matters. It's about owning fewer things but cherishing each one more deeply. By stripping away the excess, you create room for clarity, intention, and peace. Whether it\u2019s reducing physical possessions or streamlining your daily habits, minimalism can bring a sense of freedom, allowing you to invest time and energy in experiences and relationships that bring genuine joy."
                                    }
                                }
                            ]
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "input_variables",
                    "output_schema",
                    "assert",
                    "test_cases"
                ],
                "title": "AIFunctionRouteInput"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HttpExceptionModel": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "status": {
                        "type": "integer",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "status"
                ],
                "title": "HttpExceptionModel"
            },
            "InputVariable": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "pattern": "^[^\\s]+$",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "InputVariable"
            },
...
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Success"
                    }
                },
                "type": "object",
                "title": "SuccessResponse"
            },
...
            "TestCase-Output": {
                "properties": {
                    "vars": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Vars"
                    },
                    "assert": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Assertion"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assert"
                    }
                },
                "type": "object",
                "required": [
                    "vars",
                    "assert"
                ],
                "title": "TestCase"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}